parsing cmds & ihex works, but only some json so far
