+++ examples/inparse/in.txt
@@ -8,3 +8,6 @@ true
 [1,2,3]
 "a\x21b\nc\u0123d"
 {1:2,3:4}
+[1,[2,3],4]
+[1,"abc",3]
+[1,{2:3,4:5},6,[],7,{},8,[[]],9]
